requirements.txt
# Core dependencies
pygame>=2.5.0
python-osc>=1.8.0
pydub>=0.25.1
matplotlib>=3.7.0
numpy>=1.24.0
Pillow>=10.0.0
pyserial>=3.5
sacn>=1.4.0
psutil>=5.9.0

# Web framework and async support
aiohttp>=3.8.0
aiohttp-cors>=0.7.0

# Audio processing
librosa>=0.10.0

# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-playwright>=0.4.4
playwright>=1.40.0
httpx>=0.24.0